refactor: single Select component, demos moved to widgets
