+++ src/features/Select/index.ts
@@ -1,0 +1,1 @@
+export { Select } from './ui/Select';
