+++ src/features/Select/ui/Select.tsx
@@ -1,12 +1,16 @@
-import { useCallback, useState } from 'react';
-import styles from './SelectBasic.module.scss';
+import styles from './Select.module.scss';
 import ShevronIcon from '@/shared/assets/icons/chevron-down.svg?react';
 import { InputBox, InputBoxKits } from '@/shared/ui/InputBox/InputBox';
 import { Input } from '@/shared/ui/Input/Input';
 import { Button, ButtonKits } from '@/shared/ui/Button/Button';
 import { InputLabel } from '@/shared/ui/InputLabel/InputLabel';
-import { Dropdown, DropdownKits, DropdownProps } from '@/shared/ui/Dropdown/Dropdown';
+import {
+  DropdownDefault,
+  DropdownKits,
+  DropdownProps,
+} from '@/shared/ui/DropdownDefault/DropdownDefault';
 import { OptionDefault, SelectBasicOption } from '@/shared/ui/OptionDefault/OptionDefault';
+import { useSelect } from '../hooks/useSelect';
 
 type Props = {
   value: SelectBasicOption;
@@ -16,28 +20,11 @@ type Props = {
   CustomOption?: React.ComponentType;
 };
 
-export function SelectBasic({ value, options, onChange, CustomDropdown, CustomOption }: Props) {
-  const DropdownComponent = CustomDropdown || Dropdown;
-  const OptionComponent = CustomOption || OptionDefault;
+export function Select({ value, options, onChange, CustomDropdown, CustomOption }: Props) {
+  const Dropdown = CustomDropdown || DropdownDefault;
+  const Option = CustomOption || OptionDefault;
 
-  const [isOpen, setIsOpen] = useState(false);
-
-  const toggleDropdown = useCallback(() => {
-    setIsOpen(!isOpen);
-  }, [isOpen]);
-
-  const closeDropdown = useCallback(() => {
-    setIsOpen(false);
-  }, []);
-
-  const selectOption = useCallback(
-    (option: SelectBasicOption) => (e: React.MouseEvent<HTMLLIElement, MouseEvent>) => {
-      e.stopPropagation();
-      onChange(option);
-      closeDropdown();
-    },
-    [onChange, closeDropdown]
-  );
+  const { isOpen, toggleDropdown, closeDropdown, selectOption } = useSelect({ onChange });
 
   return (
     <section className={styles.section}>
@@ -52,11 +39,11 @@ export function SelectBasic({ value, options, onChange, CustomDropdown, CustomOp
         </Button>
       </InputBox>
 
-      <DropdownComponent kit={DropdownKits.SINGLE_SELECT} isOpen={isOpen}>
+      <Dropdown kit={DropdownKits.SINGLE_SELECT} isOpen={isOpen}>
         {options.map((option) => (
-          <OptionComponent key={option.value} option={option} onClick={selectOption(option)} />
+          <Option key={option.value} option={option} onClick={selectOption(option)} />
         ))}
-      </DropdownComponent>
+      </Dropdown>
     </section>
   );
 }
