+++ src/shared/ui/DropdownDefault/DropdownDefault.module.scss
