--- src/widgets/SelectBasicDemo/index.ts
@@ -1,1 +1,0 @@
-export { SelectBasicDemo } from './ui/SelectBasicDemo';
